big refactor
