+++ cmd/gogoodwe/constants/constants.go
