+++ cmd/gogoodwe/inverter/inverter.go
@@ -9,24 +9,24 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/AaronSaikovski/gogoodwe/constants"
-	"github.com/AaronSaikovski/gogoodwe/types"
-	"github.com/AaronSaikovski/gogoodwe/utils"
+	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/constants"
+	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/types"
+	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/utils"
 )
 
 // fetchInverterData - Fetches Data from the Inverter via the specified PowerstationID using theSEMs API
 func fetchInverterData(SemsResponseData *types.LoginResponse, UserLogin *types.LoginCredentials, PowerstationOutputData *types.InverterData) error {
 
 	// get the Token header data
-	tokenMapJSONData, tokenMapJSONErr := dataTokenJSON(SemsResponseData)
-	if tokenMapJSONErr != nil {
-		return tokenMapJSONErr
+	tokenMapJSONData, err := utils.DataTokenJSON(SemsResponseData)
+	if err != nil {
+		return err
 	}
 
 	// get the Powerstation ID header data
-	powerStationMapJSONData, powerStationMapJSONErr := powerStationIDJSON(UserLogin)
-	if powerStationMapJSONErr != nil {
-		return powerStationMapJSONErr
+	powerStationMapJSONData, err := utils.PowerStationIDJSON(UserLogin)
+	if err != nil {
+		return err
 	}
 
 	//Get the url from the Auth API and append the data url part
@@ -39,7 +39,7 @@ func fetchInverterData(SemsResponseData *types.LoginResponse, UserLogin *types.L
 	}
 
 	//Add headers pass in the pointer to set the headers on the request object
-	setHeaders(req, tokenMapJSONData)
+	utils.SetHeaders(req, tokenMapJSONData)
 
 	//make the API Call
 	client := &http.Client{Timeout: constants.HTTPTimeout * time.Second}
@@ -52,9 +52,9 @@ func fetchInverterData(SemsResponseData *types.LoginResponse, UserLogin *types.L
 	defer resp.Body.Close()
 
 	// Get the response body
-	respBody, respBodyErr := utils.FetchResponseBody(resp.Body)
-	if respBodyErr != nil {
-		return respBodyErr
+	respBody, err := utils.FetchResponseBody(resp.Body)
+	if err != nil {
+		return err
 	}
 
 	//marshall response to SemsRespInfo struct
