+++ cmd/gogoodwe/types/inverterrdata.go
