+++ cmd/gogoodwe/types/logincredentials.go
