+++ cmd/gogoodwe/types/loginresponse.go
