+++ cmd/gogoodwe/utils/args.go
@@ -1,6 +1,6 @@
 package utils
 
-import "github.com/AaronSaikovski/gogoodwe/constants"
+import "github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/constants"
 
 // Args - struct using go-arg- https://github.com/alexflint/go-arg
 type Args struct {
