+++ cmd/gogoodwe/utils/jsonutils.go
@@ -2,25 +2,17 @@
 # Name: powerstationhelper - helper functions to get the Powerstation Data from the API
 # Author: Aaron Saikovski - asaikovski@outlook.com
 */
-package inverter
+package utils
 
 import (
 	"encoding/json"
-	"net/http"
 	"strconv"
 
-	"github.com/AaronSaikovski/gogoodwe/types"
-	"github.com/AaronSaikovski/gogoodwe/utils"
+	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/types"
 )
 
-// setHeaders - Set the headers for the SEMS Data API
-func setHeaders(r *http.Request, tokenstring []byte) {
-	r.Header.Add("Content-Type", "application/json")
-	r.Header.Add("Token", string(tokenstring))
-}
-
 // DataTokenJSON - Makes a map for the token to be passed to the Data API header and returns a JSON string
-func dataTokenJSON(SemsResponseData *types.LoginResponse) ([]byte, error) {
+func DataTokenJSON(SemsResponseData *types.LoginResponse) ([]byte, error) {
 	tokenMap := make(map[string]string)
 	tokenMap["version"] = "v2.1.0"
 	tokenMap["client"] = "ios"
@@ -35,7 +27,7 @@ func dataTokenJSON(SemsResponseData *types.LoginResponse) ([]byte, error) {
 }
 
 // PowerStationIDJSON - Makes a map for the powerStationId to be passed to the Data API header and returns a JSON string
-func powerStationIDJSON(UserLogin *types.LoginCredentials) ([]byte, error) {
+func PowerStationIDJSON(UserLogin *types.LoginCredentials) ([]byte, error) {
 	powerStationMap := make(map[string]string)
 	powerStationMap["powerStationId"] = UserLogin.PowerStationID
 
@@ -48,6 +40,6 @@ func powerStationIDJSON(UserLogin *types.LoginCredentials) ([]byte, error) {
 func GetDataJSON(PowerstationOutputData *types.InverterData) ([]byte, error) {
 
 	// Get the response and return any errors
-	resp, err := utils.MarshalStructToJSON(&PowerstationOutputData)
+	resp, err := MarshalStructToJSON(&PowerstationOutputData)
 	return resp, err
 }
