+++ cmd/gogoodwe/utils/paramcheck.go
