+++ cmd/gogoodwe/utils/responsehandler.go
